# Improve performance of digests