--- a/digest/src/blake2/mod.rs
+++ b/digest/src/blake2/mod.rs
@@ -15,10 +15,11 @@
 //!
 //! [md]: https://en.wikipedia.org/wiki/Merkle%E2%80%93Damg%C3%A5rd_construction
 
+#![allow(doc_markdown)]
+
 use core::marker::PhantomData;
-use core::{mem, ptr};
-use core::num::Wrapping as W;
-use core::ops::{Shl, BitXor, Mul};
+use core::ptr;
+use core::ops::Mul;
 
 use static_buffer::{FixedBuffer64, FixedBuffer128, FixedBuf};
 use byteorder::{ByteOrder, LittleEndian};
@@ -27,23 +28,24 @@ use typenum::uint::Unsigned;
 use typenum::consts::{U8, U16, U20, U28, U32, U48, U64, U128};
 
 use Digest;
+use wrapping::*;
 
-const BLAKE2S_INIT: [W<u32>; 8] = [W(0x6a09e667),
-                                   W(0xbb67ae85),
-                                   W(0x3c6ef372),
-                                   W(0xa54ff53a),
-                                   W(0x510e527f),
-                                   W(0x9b05688c),
-                                   W(0x1f83d9ab),
-                                   W(0x5be0cd19)];
-const BLAKE2B_INIT: [W<u64>; 8] = [W(0x6a09e667f3bcc908),
-                                   W(0xbb67ae8584caa73b),
-                                   W(0x3c6ef372fe94f82b),
-                                   W(0xa54ff53a5f1d36f1),
-                                   W(0x510e527fade682d1),
-                                   W(0x9b05688c2b3e6c1f),
-                                   W(0x1f83d9abfb41bd6b),
-                                   W(0x5be0cd19137e2179)];
+const BLAKE2S_INIT: [w32; 8] = [W(0x6a09e667),
+                                W(0xbb67ae85),
+                                W(0x3c6ef372),
+                                W(0xa54ff53a),
+                                W(0x510e527f),
+                                W(0x9b05688c),
+                                W(0x1f83d9ab),
+                                W(0x5be0cd19)];
+const BLAKE2B_INIT: [w64; 8] = [W(0x6a09e667f3bcc908),
+                                W(0xbb67ae8584caa73b),
+                                W(0x3c6ef372fe94f82b),
+                                W(0xa54ff53a5f1d36f1),
+                                W(0x510e527fade682d1),
+                                W(0x9b05688c2b3e6c1f),
+                                W(0x1f83d9abfb41bd6b),
+                                W(0x5be0cd19137e2179)];
 
 const SIGMA: [[usize; 16]; 12] = [[0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15],
                                   [14, 10, 4, 8, 9, 15, 13, 6, 1, 12, 0, 2, 11, 7, 5, 3],
@@ -60,14 +62,14 @@ const SIGMA: [[usize; 16]; 12] = [[0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13,
 
 macro_rules! G {
     ($v:ident, $a:expr, $b:expr, $c:expr, $d:expr, $x:expr, $y:expr) => {
-        $v[$a] = $v[$a] + $v[$b] + $x;
-        $v[$d] = W(($v[$d] ^ $v[$a]).0.rotate_right(R1));
-        $v[$c] = $v[$c] + $v[$d];
-        $v[$b] = W(($v[$b] ^ $v[$c]).0.rotate_right(R2));
-        $v[$a] = $v[$a] + $v[$b] + $y;
-        $v[$d] = W(($v[$d] ^ $v[$a]).0.rotate_right(R3));
-        $v[$c] = $v[$c] + $v[$d];
-        $v[$b] = W(($v[$b] ^ $v[$c]).0.rotate_right(R4));
+        $v[$a] +=           $v[$b] + $x;
+        $v[$d]  = ($v[$d] ^ $v[$a]).rotate_right(R1);
+        $v[$c] +=           $v[$d];
+        $v[$b]  = ($v[$b] ^ $v[$c]).rotate_right(R2);
+        $v[$a] +=           $v[$b] + $y;
+        $v[$d]  = ($v[$d] ^ $v[$a]).rotate_right(R3);
+        $v[$c] +=           $v[$d];
+        $v[$b]  = ($v[$b] ^ $v[$c]).rotate_right(R4);
     }
 }
 
@@ -76,59 +78,57 @@ struct State<Word: Copy> {
     h: [W<Word>; 8],
 }
 
-impl<Word> State<Word>
-    where Word: From<u8> + From<u32> + Copy + BitXor<Output = Word> + Shl<usize, Output = Word> + Default
-{
-    fn new(mut state: [W<Word>; 8], key_size: u8, size: u8) -> Self {
-        state[0] = W(state[0].0 ^ 0x1010000u32.into() ^ ((key_size as u32) << 8).into() ^ size.into());
-
-        State { h: state }
-    }
-}
-
 macro_rules! blake2_state {
     ($word:ty, $read:ident, $init:ident, $rounds:expr, $r1:expr, $r2:expr, $r3:expr, $r4:expr) => {
         impl State<$word> {
+            fn new(key_size: u8, size: u8) -> Self {
+                let mut state = $init;
+
+                state[0] ^= W(0x01010000 ^ ((key_size as $word) << 8) ^ (size as $word));
+
+                State { h: state }
+            }
+
+            #[inline(always)]
             fn compress(&mut self, input: &[u8], len: Length<$word>, last: bool) {
+                debug_assert!(input.len() % 16 == 0);
                 const R1: u32 = $r1;
                 const R2: u32 = $r2;
                 const R3: u32 = $r3;
                 const R4: u32 = $r4;
 
-                let word_bytes = mem::size_of::<$word>();
-
                 let mut message = [W(0); 16];
-                for (word, chunk) in message.iter_mut().zip(input.chunks(word_bytes)) {
+                for (word, chunk) in message.iter_mut().zip(input.chunks(input.len() / 16)) {
                     *word = W(LittleEndian::$read(chunk));
                 }
 
                 let mut v = [W(0); 16];
-                for (v, state) in v.iter_mut().take(8).zip(&self.h) {
+                for (v, state) in v.iter_mut().zip(self.h.iter().chain(&$init)) {
                     *v = *state;
                 }
-                for (v, iv) in v.iter_mut().skip(8).zip(&$init) {
-                    *v = *iv;
-                }
-                v[12].0 ^= len.0;
-                v[13].0 ^= len.1;
+                v[12].0 ^= len.0.to_le();
+                v[13].0 ^= len.1.to_le();
                 if last {
                     v[14] = !v[14];
                 }
 
                 for sigma in &SIGMA[..$rounds] {
-                    G!(v, 0, 4, 8,  12, message[sigma[0]],  message[sigma[1]]);
-                    G!(v, 1, 5, 9,  13, message[sigma[2]],  message[sigma[3]]);
+                    G!(v, 0, 4,  8, 12, message[sigma[0]],  message[sigma[1]]);
+                    G!(v, 1, 5,  9, 13, message[sigma[2]],  message[sigma[3]]);
                     G!(v, 2, 6, 10, 14, message[sigma[4]],  message[sigma[5]]);
                     G!(v, 3, 7, 11, 15, message[sigma[6]],  message[sigma[7]]);
 
                     G!(v, 0, 5, 10, 15, message[sigma[8]],  message[sigma[9]]);
                     G!(v, 1, 6, 11, 12, message[sigma[10]], message[sigma[11]]);
-                    G!(v, 2, 7, 8,  13, message[sigma[12]], message[sigma[13]]);
-                    G!(v, 3, 4, 9,  14, message[sigma[14]], message[sigma[15]]);
+                    G!(v, 2, 7,  8, 13, message[sigma[12]], message[sigma[13]]);
+                    G!(v, 3, 4,  9, 14, message[sigma[14]], message[sigma[15]]);
                 }
 
-                for i in 0..8 {
-                    self.h[i] = self.h[i] ^ v[i] ^ v[i + 8];
+                let (head, tail) = v.split_at(8);
+                let vs = head.iter().zip(tail);
+
+                for (h, (&v1, &v2)) in self.h.iter_mut().zip(vs) {
+                    *h ^= v1 ^ v2;
                 }
             }
         }
@@ -141,21 +141,20 @@ blake2_state!(u64, read_u64, BLAKE2B_INIT, 12, 32, 24, 16, 63);
 #[derive(Copy, Clone, Debug)]
 struct Length<T>(T, T);
 
-
 impl Length<u32> {
     fn increment(&mut self, val: usize) {
-        self.0.wrapping_add(val as u32);
+        self.0 = self.0.wrapping_add(val as u32);
     }
 }
 
 impl Length<u64> {
     fn increment(&mut self, val: usize) {
-        self.0.wrapping_add(val as u64);
+        self.0 = self.0.wrapping_add(val as u64);
     }
 }
 
 macro_rules! blake2 {
-    ($(#[$attr:meta])* struct $name:ident<$word:ty>, $init:ident, $buf:ty, $bsize:ty) => {
+    ($(#[$attr:meta])* struct $name:ident<$word:ty>, $buf:ty, $bsize:ty, $wsize: expr) => {
         #[derive(Clone)]
         $(#[$attr])*
         pub struct $name<Size: Unsigned + Clone> {
@@ -165,35 +164,34 @@ macro_rules! blake2 {
             _phantom: PhantomData<Size>,
         }
 
-        impl<Size: Unsigned + Clone> $name<Size> {
-            /// Initialize BLAKE2 hash function with custom key
+        impl<Size> $name<Size>
+            where Size: Unsigned + Clone + ArrayLength<u8> + Mul<U8>,
+                  <Size as Mul<U8>>::Output: ArrayLength<u8>
+        {
+/// Default
+            pub fn default() -> Self {
+                Self::with_key(&[])
+            }
+
+/// Initialize BLAKE2 hash function with custom key
             pub fn with_key<K: AsRef<[u8]>>(key: K) -> Self {
                 let key = key.as_ref();
 
+                assert!(key.len() <= $wsize);
+
                 let mut ret = $name {
-                    state: State::new($init, key.len() as u8, Size::to_u8()),
+                    state: State::<$word>::new(key.len() as u8, Size::to_u8()),
                     len: Length(0, 0),
                     buffer: <$buf>::new(),
                     _phantom: PhantomData
                 };
 
                 if !key.is_empty() {
-                    ret.buffer.input(key, |_| {});
+                    ret.update(key);
                     ret.buffer.zero_until(<$buf>::size());
-
-                    ret.len.increment(<$buf>::size());
-                    ret.state.compress(ret.buffer.full_buffer(),
-                                       ret.len,
-                                       false);
                 }
 
                 ret
-            }
-        }
-
-        impl<Size: Unsigned + Clone> Default for $name<Size> {
-            fn default() -> Self {
-                Self::with_key(&[])
             }
         }
 
@@ -207,10 +205,11 @@ macro_rules! blake2 {
 
             fn update<T: AsRef<[u8]>>(&mut self, input: T) {
                 let input = input.as_ref();
+
                 let state = &mut self.state;
                 let len = &mut self.len;
 
-                self.buffer.input(&input, |d| {
+                self.buffer.input(input, |d| {
                     len.increment(d.len());
                     state.compress(d, *len, false);
                 })
@@ -221,17 +220,15 @@ macro_rules! blake2 {
                 self.len.increment(rest);
                 self.buffer.zero_until(<$buf>::size());
 
-                self.state.compress(self.buffer.full_buffer(),
-                                    self.len,
-                                    true);
+                self.state.compress(self.buffer.full_buffer(), self.len, true);
 
                 let mut out = out.as_mut();
                 assert!(out.len() >= Self::output_bytes());
 
                 unsafe {
                     ptr::copy_nonoverlapping(self.state.h.as_ptr() as *const u8,
-                                             out.as_mut_ptr(),
-                                             Self::output_bytes())
+                    out.as_mut_ptr(),
+                    Self::output_bytes())
                 }
             }
         }
@@ -240,11 +237,11 @@ macro_rules! blake2 {
 
 blake2! {
     /// General BLAKE2s implementation
-    struct Blake2s<u32>, BLAKE2S_INIT, FixedBuffer64, U64
+    struct Blake2s<u32>, FixedBuffer64, U64, 32
 }
 blake2! {
     /// General BLAKE2b implementation
-    struct Blake2b<u64>, BLAKE2B_INIT, FixedBuffer128, U128
+    struct Blake2b<u64>, FixedBuffer128, U128, 64
 }
 
 /// BLAKE2s-128 implementation