--- a/tests/mac/mod.rs
+++ b/tests/mac/mod.rs
@@ -1,1 +1,0 @@
-mod hmac;