… .travis.yml
… Cargo.toml
… digest/Cargo.toml
… digest/src/sha1.rs
… digest/src/sha1/intrinsics.rs
… digest/src/sha1/mod.rs
… digest/src/sha2.rs
… digest/src/simd.rs
… digest/src/tiger/mod.rs
… digest/src/whirlpool/mod.rs
… digest/src/whirlpool/sboxes.rs
… digest/src/wrapping.rs
… digest/tests/blake2/blake2-kat.h
… digest/tests/blake2/blake2-kat.json
… digest/tests/blake2/blake2b.toml
… digest/tests/blake2/blake2bp-kat.txt
… digest/tests/blake2/blake2s-kat.txt
… digest/tests/blake2/blake2s.toml
… digest/tests/blake2/blake2sp-kat.txt
… digest/tests/blake2/mod.rs
… digest/tests/md5/mod.rs
… digest/tests/md5/rfc1321.toml
… digest/tests/sha3/mod.rs
… digest/tests/sha3/sha224.toml
… digest/tests/sha3/sha256.toml
… digest/tests/sha3/sha384.toml
… digest/tests/sha3/sha512.toml
… digest/tests/utils/mod.rs
… digest/tests/utils/tests.rs
… digest/tests/whirlpool/iso.toml