```
 name                               before ns/iter     after ns/iter      diff ns/iter   diff % 
 blake2::blake2b::_1024_block_size  4,401 (232 MB/s)   3,857 (265 MB/s)           -544  -12.36% 
 blake2::blake2b::_16_block_size    79 (202 MB/s)      67 (238 MB/s)               -12  -15.19% 
 blake2::blake2b::_256_block_size   1,298 (197 MB/s)   970 (263 MB/s)             -328  -25.27% 
 blake2::blake2b::_64_block_size    286 (223 MB/s)     246 (260 MB/s)              -40  -13.99% 
 blake2::blake2b::_8192_block_size  36,314 (225 MB/s)  30,771 (266 MB/s)        -5,543  -15.26% 
 blake2::blake2s::_1024_block_size  7,850 (130 MB/s)   6,368 (160 MB/s)         -1,482  -18.88% 
 blake2::blake2s::_16_block_size    130 (123 MB/s)     104 (153 MB/s)              -26  -20.00% 
 blake2::blake2s::_256_block_size   1,949 (131 MB/s)   1,601 (159 MB/s)           -348  -17.86% 
 blake2::blake2s::_64_block_size    505 (126 MB/s)     403 (158 MB/s)             -102  -20.20% 
 blake2::blake2s::_8192_block_size  61,197 (133 MB/s)  50,412 (162 MB/s)       -10,785  -17.62% 
 md5::_1024_block_size              3,353 (305 MB/s)   2,758 (371 MB/s)           -595  -17.75% 
 md5::_256_block_size               860 (297 MB/s)     693 (369 MB/s)             -167  -19.42% 
 md5::_64_block_size                214 (299 MB/s)     174 (367 MB/s)              -40  -18.69% 
 md5::_8192_block_size              26,523 (308 MB/s)  22,079 (371 MB/s)        -4,444  -16.76% 
 ripemd::_256_block_size            1,888 (135 MB/s)   1,623 (157 MB/s)           -265  -14.04% 
 sha2::sha224::_1024_block_size     8,397 (121 MB/s)   6,970 (146 MB/s)         -1,427  -16.99% 
 sha2::sha224::_16_block_size       142 (112 MB/s)     115 (139 MB/s)              -27  -19.01% 
 sha2::sha224::_256_block_size      2,186 (117 MB/s)   1,750 (146 MB/s)           -436  -19.95% 
 sha2::sha256::_1024_block_size     8,411 (121 MB/s)   6,844 (149 MB/s)         -1,567  -18.63% 
 sha2::sha384::_1024_block_size     4,716 (217 MB/s)   4,022 (254 MB/s)           -694  -14.72% 
 sha2::sha384::_8192_block_size     37,303 (219 MB/s)  32,096 (255 MB/s)        -5,207  -13.96% 
 sha2::sha512::_16_block_size       116 (137 MB/s)     70 (228 MB/s)               -46  -39.66% 
 sha3::sha224::_256_block_size      1,446 (177 MB/s)   1,207 (212 MB/s)           -239  -16.53% 
 sha3::sha224::_64_block_size       396 (161 MB/s)     309 (207 MB/s)              -87  -21.97% 
 sha3::sha224::_8192_block_size     46,358 (176 MB/s)  37,727 (217 MB/s)        -8,631  -18.62% 
 sha3::sha256::_16_block_size       95 (168 MB/s)      82 (195 MB/s)               -13  -13.68% 
 sha3::sha256::_8192_block_size     49,977 (163 MB/s)  39,906 (205 MB/s)       -10,071  -20.15% 
 sha3::sha384::_1024_block_size     8,573 (119 MB/s)   6,479 (158 MB/s)         -2,094  -24.43% 
 sha3::sha384::_16_block_size       139 (115 MB/s)     106 (150 MB/s)              -33  -23.74% 
 sha3::sha384::_256_block_size      2,654 (96 MB/s)    1,633 (156 MB/s)         -1,021  -38.47% 
 sha3::sha384::_64_block_size       598 (107 MB/s)     444 (144 MB/s)             -154  -25.75% 
 sha3::sha384::_8192_block_size     65,292 (125 MB/s)  51,886 (157 MB/s)       -13,406  -20.53% 
 sha3::sha512::_16_block_size       182 (87 MB/s)      149 (107 MB/s)              -33  -18.13% 
 sha3::sha512::_256_block_size      2,908 (88 MB/s)    2,345 (109 MB/s)           -563  -19.36% 
 tiger::tiger2::_16_block_size      54 (296 MB/s)      46 (347 MB/s)                -8  -14.81% 
 tiger::tiger2::_256_block_size     770 (332 MB/s)     620 (412 MB/s)             -150  -19.48% 
 tiger::tiger2::_64_block_size      302 (211 MB/s)     159 (402 MB/s)             -143  -47.35% 
 tiger::tiger2::_8192_block_size    25,652 (319 MB/s)  19,915 (411 MB/s)        -5,737  -22.36% 
 tiger::tiger::_1024_block_size     2,832 (361 MB/s)   2,475 (413 MB/s)           -357  -12.61% 
 tiger::tiger::_16_block_size       53 (301 MB/s)      44 (363 MB/s)                -9  -16.98% 
 tiger::tiger::_64_block_size       197 (324 MB/s)     160 (400 MB/s)              -37  -18.78% 
```