@homu 2016-09-17
:pushpin: Commit 39c2b17 has been approved by `hauleth`

<!-- @homu r=hauleth 39c2b17309a09bfacc916d67df8c648d566680fb -->